refactor: extract ModalFormData/Val types to break source-tree cycle
The previous commit had FormResult.ts importing executeTransformation
from templateParser.ts, while templateParser.ts already imported the
ModalFormData and Val types back from FormResult.ts — a source-level
cycle that, while currently benign (the reverse leg was type-only),
would silently break if a future change made templateParser.ts import
a runtime value from FormResult.ts.

Move Val, ModalFormData, isPrimitive, and isPrimitiveArray into a new
src/core/formResultTypes.ts. FormResult.ts re-exports them so external
consumers (FormModal, FormResult.test, etc.) keep working unchanged.
templateParser.ts and formDataFromFormDefaults.ts now import from the
new file directly, eliminating the cycle entirely.

Addresses Greptile review comment on PR #482.
